debt: Transformed mediaBackup into a real module
