+++ src/drive/mobile/ducks/mediaBackup/components/UploadPreparing.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import classnames from 'classnames'
 import PropTypes from 'prop-types'
-import styles from './styles'
+import styles from '../styles'
 
 const UploadPreparing = ({ t }) => (
   <div className={styles['coz-upload-status-wrapper']}>
