+++ src/drive/mobile/ducks/mediaBackup/components/UploadQuotaError.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import classnames from 'classnames'
 import PropTypes from 'prop-types'
-import styles from './styles'
+import styles from '../styles'
 
 const UploadQuotaError = ({ t, url }) => (
   <div
